:tada: Hello, Void Linux!
For all intents and purposes, this is a complete rewrite of the old dotfiles. Some amateur mistakes such as *pushing the entirety of `.oh-my-zsh` and `.emacs.d`* have been avoided (thankfully).

Vim folding to neatly organize longer configs, so using Vim (or NeoVim) is the preferred option of viewing the configs unless some other editor supports Vim folding.

Notably, `config/sxhkd` includes a README with the keyboard combinations shown in a table. The goal is to have each config folder include a README for documentation purposes.